pyproject.toml
[project]
name = "less-reddit"
version = "0.1.0"
description = "A terminal-based Reddit browser with a less-like scrollable viewer."
authors = [
    { name="Alicu96", email="trihm1996@gmail.com" }
]
dependencies = [
    "praw",
    "click",
    "python-dotenv"
]
requires-python = ">=3.7"

[project.scripts]
less-reddit = "less_reddit.cli:main"